Extract typescript rules
