+++ index.js
@@ -1,17 +1,5 @@
 "use strict";
 
-const xoTypeScriptConfig = require("eslint-config-xo-typescript");
-
-function customize(config, rule, customizer) {
-	const [type, ruleConfig] = config.rules[rule];
-	// Spread shallow-clones the object
-	const newRuleConfig = { ...ruleConfig };
-	customizer(newRuleConfig);
-	return {
-		[rule]: [type, newRuleConfig],
-	};
-}
-
 const config = {
 	env: {
 		browser: true,
@@ -19,10 +7,9 @@ const config = {
 	ignorePatterns: [".idea", "dist", "**/__mocks__/**"],
 	plugins: ["filenames", "jsx-a11y"],
 	extends: [
-		"./xo-plugins-config.js",
-
 		"xo", // Full config: https://github.com/xojs/eslint-config-xo/blob/main/index.js
-		"xo-typescript", // Full config: https://github.com/xojs/eslint-config-xo-typescript/blob/main/index.js
+		"./xo-plugins-config.js", // Vendored from xojs/xo package, it must be here as a baseline
+
 		"prettier", // Disable style-related rules
 		"plugin:security/recommended-legacy",
 		"plugin:unicorn/recommended", // Full config: https://github.com/sindresorhus/eslint-plugin-unicorn/blob/main/configs/recommended.js
@@ -32,6 +19,7 @@ const config = {
 		"./plugins/jsdoc.js",
 		"./plugins/react.js",
 		"./plugins/import.js",
+		"./plugins/typescript.js",
 
 		/**************************************************************
 		 * Only add test rules and plugins to the "./tests.js" config *
@@ -41,9 +29,7 @@ const config = {
 		// Enable extra rules
 
 		"no-restricted-imports": ["error", require("./no-restricted-imports")],
-
 		"no-restricted-syntax": ["error", ...require("./no-restricted-syntax")],
-
 		"no-mixed-operators": [
 			"error",
 			{
@@ -62,38 +48,6 @@ const config = {
 
 		// Customize some rules
 		quotes: ["error", "double", { avoidEscape: true }], // Matches Prettier, but also replaces backticks
-
-		...customize(
-			xoTypeScriptConfig,
-			"@typescript-eslint/ban-types",
-			(config) => {
-				delete config.types.null;
-			}
-		),
-
-		// Reason: https://github.com/pixiebrix/pixiebrix-extension/pull/7703
-		"@typescript-eslint/restrict-template-expressions": [
-			"error",
-			{ allowNever: true, allowNumber: true },
-		],
-
-		// We want to have a default case to check for `never`
-		"@typescript-eslint/switch-exhaustiveness-check": [
-			"error",
-			{
-				allowDefaultCaseForExhaustiveSwitch: true,
-				requireDefaultForNonUnion: true,
-			},
-		],
-
-		"@typescript-eslint/no-non-null-assertion": "error",
-		"@typescript-eslint/no-explicit-any": [
-			"error",
-			{
-				fixToUnknown: true,
-				ignoreRestArgs: true,
-			},
-		],
 
 		"unicorn/prefer-export-from": [
 			"error",
@@ -172,49 +126,15 @@ const config = {
 		"unicorn/no-nested-ternary": "off", // Sometimes it conflicts with Prettier
 		"unicorn/prefer-set-has": "off", // Not always worth the extra code
 		"unicorn/prefer-top-level-await": "off", // No advantage in browsers
-		"@typescript-eslint/triple-slash-reference": "off", // No alternative sometimes
-		"@typescript-eslint/consistent-type-definitions": "off", // `type` cannot be used to extend globals
-		"@typescript-eslint/no-dynamic-delete": "off", // Already covered by `security/detect-object-injection`
-		"@typescript-eslint/consistent-type-assertions": "off", // Our current typing has too many `unknowns` for this to be applicable https://github.com/typescript-eslint/typescript-eslint/issues/4462
-
-		// Disable rule until we find a better config https://github.com/pixiebrix/eslint-config-pixiebrix/issues/5
-		"@typescript-eslint/naming-convention": "off",
-
-		// Requires strictNullChecks
-		"@typescript-eslint/prefer-nullish-coalescing": "off",
 
 		// Maybe later, opinionated
 		"unicorn/prefer-ternary": "off",
-		"@typescript-eslint/member-ordering": "off",
-		"@typescript-eslint/no-empty-function": "off",
 
 		"node/file-extension-in-import": "off",
 		"node/prefer-global/process": "off", // `process.env` is required by webpack
 		"node/prefer-global/buffer": "off",
-
-		"@typescript-eslint/no-use-before-define": [
-			"error",
-			{
-				// Disabling functions -- functions are hoisted and not a risk
-				// https://eslint.org/docs/latest/rules/no-use-before-define#options
-				functions: false,
-				// https://typescript-eslint.io/rules/no-use-before-define/#options
-				ignoreTypeReferences: false,
-			},
-		],
 	},
 	overrides: [
-		{
-			// JS files shouldn't have TypeScript rules, but it's bothersome to separate them properly
-			files: ["**/*.js"],
-			rules: {
-				"@typescript-eslint/no-unsafe-argument": "off",
-				"@typescript-eslint/no-unsafe-assignment": "off",
-				"@typescript-eslint/no-unsafe-call": "off",
-				"@typescript-eslint/no-unsafe-member-access": "off",
-				"@typescript-eslint/no-unsafe-return": "off",
-			},
-		},
 		{
 			files: ["**/*.tsx", "**/use*.ts"],
 			excludedFiles: ["*.test.tsx", "*.stories.tsx"],
